third_party/WebKit/LayoutTests/html/tabular_data/tr_exceptions.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../resources/js-test.js"></script>
</head>
<body>
    <script>
        description("This test should trigger exceptions on HTMLTableRowElement, and verify that the messages are reasonably helpful.");

        var tr = document.createElement('tr');

        shouldThrow("tr.insertCell(-2)");
        shouldThrow("tr.insertCell(1)");

        tr.insertCell();

        shouldThrow("tr.deleteCell(-2)");
        shouldThrow("tr.deleteCell(2)");
        shouldThrow("tr.deleteCell()");
    </script>
</body>
</html>
