third_party/blink/web_tests/platform/generic/html/tabular_data/tr_exceptions-expected.txt

This test should trigger exceptions on HTMLTableRowElement, and verify that the messages are reasonably helpful.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS tr.insertCell(-2) threw exception IndexSizeError: Failed to execute 'insertCell' on 'HTMLTableRowElement': The value provided (-2) is outside the range [-1, 0]..
PASS tr.insertCell(1) threw exception IndexSizeError: Failed to execute 'insertCell' on 'HTMLTableRowElement': The value provided (1) is outside the range [-1, 0]..
PASS tr.deleteCell(-2) threw exception IndexSizeError: Failed to execute 'deleteCell' on 'HTMLTableRowElement': The value provided (-2) is outside the range [0, 1)..
PASS tr.deleteCell(2) threw exception IndexSizeError: Failed to execute 'deleteCell' on 'HTMLTableRowElement': The value provided (2) is outside the range [0, 1)..
PASS tr.deleteCell() threw exception TypeError: Failed to execute 'deleteCell' on 'HTMLTableRowElement': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
